#7263 replace DashboardCriteriaInterval with DashboardCriteria
