+++ sormas-api/src/main/java/de/symeda/sormas/api/dashboard/DashboardCriteria.java
@@ -17,6 +17,8 @@ public class DashboardCriteria extends BaseCriteria implements Serializable {
 	private CriteriaDateType newCaseDateType;
 	private Date dateFrom;
 	private Date dateTo;
+	private Date previousDateFrom;
+	private Date previousDateTo;
 
 	private boolean includeNotACaseClassification;
 
@@ -77,5 +79,13 @@ public class DashboardCriteria extends BaseCriteria implements Serializable {
 	public DashboardCriteria includeNotACaseClassification(boolean includeNotACaseClassification) {
 		this.includeNotACaseClassification = includeNotACaseClassification;
 		return this;
+	}
+
+	public Date getPreviousDateFrom() {
+		return previousDateFrom;
+	}
+
+	public Date getPreviousDateTo() {
+		return previousDateTo;
 	}
 }
