+++ sormas-api/src/main/java/de/symeda/sormas/api/dashboard/DashboardFacade.java
@@ -7,13 +7,13 @@ import java.util.Map;
 import javax.ejb.Remote;
 
 import de.symeda.sormas.api.caze.CaseClassification;
-import de.symeda.sormas.api.caze.NewCaseDateType;
 import de.symeda.sormas.api.disease.DiseaseBurdenDto;
 import de.symeda.sormas.api.event.EventStatus;
 import de.symeda.sormas.api.infrastructure.district.DistrictReferenceDto;
 import de.symeda.sormas.api.infrastructure.region.RegionReferenceDto;
 import de.symeda.sormas.api.person.PresentCondition;
 import de.symeda.sormas.api.sample.PathogenTestResultType;
+import de.symeda.sormas.api.utils.criteria.CriteriaDateType;
 
 @Remote
 public interface DashboardFacade {
@@ -41,5 +41,5 @@ public interface DashboardFacade {
 		Date toDate,
 		Date previousFromDate,
 		Date previousToDate,
-		NewCaseDateType newCaseDateType);
+		CriteriaDateType newCaseDateType);
 }
